feat(execution): align remote callback governance
